+++ src/components/Footer.tsx
@@ -1,6 +1,6 @@
 import { FooterLogoImage } from "@/assets/images/logo-images";
 import { CAMP_LIST, INFO_LIST, SNS_LIST } from "@/constants";
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib";
 
 interface FooterProps {
   contentContainerMaxWidth?: number;
