+++ src/components/common/Button.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib";
 import type { ComponentProps } from "react";
 import { cva, type VariantProps } from "class-variance-authority";
