+++ src/components/common/LoadingUi.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib";
 
 interface LoadingUiProps {
   dotSize?: number;
