+++ src/components/common/SideBarTapButton.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib";
 import type { ComponentProps } from "react";
 
 function SideBarTapButton({
