+++ src/components/common/dropdown/index.ts
@@ -4,25 +4,10 @@ import DropdownContent from "@/components/common/dropdown/DropdownContent";
 import DropdownItem from "@/components/common/dropdown/DropdownItem";
 import DropdownCustomInput from "@/components/common/dropdown/DropdownCustomInput";
 
-import {
-  DEFAULT_PLACEHOLDER,
-  CUSTOM_OPTION,
-  MAX_CUSTOM_INPUT_LENGTH,
-} from "@/components/common/dropdown/constants";
-
-import type {
-  DropdownOption,
-  SelectedOption,
-} from "@/components/common/dropdown/types";
-
 export {
   Dropdown,
   DropdownTrigger,
   DropdownContent,
   DropdownItem,
   DropdownCustomInput,
-  DEFAULT_PLACEHOLDER,
-  CUSTOM_OPTION,
-  MAX_CUSTOM_INPUT_LENGTH,
 };
-export type { DropdownOption, SelectedOption };
