+++ src/components/common/index.ts
@@ -1,0 +1,17 @@
+import { Dropdown } from "@/components/common/dropdown";
+import Button from "@/components/common/Button";
+import Input from "@/components/common/Input";
+import LoadingUi from "@/components/common/LoadingUi";
+import Password from "@/components/common/Password";
+import SideBarTapButton from "@/components/common/SideBarTapButton";
+import Textarea from "@/components/common/Textarea";
+
+export {
+  Dropdown,
+  Button,
+  Input,
+  LoadingUi,
+  Password,
+  SideBarTapButton,
+  Textarea,
+};
