+++ src/components/common/toast/Toast.tsx
@@ -1,4 +1,4 @@
-import type { Toast } from "@/types/toast-types";
+import type { Toast } from "@/types";
 import BigToast from "@/components/common/toast/BigToast";
 import SmallToast from "@/components/common/toast/SmallToast";
