+++ src/components/index.tsx
@@ -1,3 +1,4 @@
 import Footer from "@/components/Footer";
+import MainContentSelector from "@/components/MainContentSelector";
 
-export { Footer };
+export { Footer, MainContentSelector };
