+++ src/components/layout/index.ts
@@ -1,0 +1,3 @@
+import RootLayout from "@/components/layout/RootLayout";
+
+export { RootLayout };
