+++ src/constants/index.ts
@@ -1,4 +1,17 @@
 import { CAMP_LIST, INFO_LIST, SNS_LIST } from "@/constants/footer-constants";
 import { MODAL_ANIMATION_TIME_MS } from "@/constants/modal-constants";
+import {
+  DEFAULT_PLACEHOLDER,
+  CUSTOM_OPTION,
+  MAX_CUSTOM_INPUT_LENGTH,
+} from "@/constants/dropdown-constants";
 
-export { CAMP_LIST, INFO_LIST, SNS_LIST, MODAL_ANIMATION_TIME_MS };
+export {
+  CAMP_LIST,
+  INFO_LIST,
+  SNS_LIST,
+  MODAL_ANIMATION_TIME_MS,
+  DEFAULT_PLACEHOLDER,
+  CUSTOM_OPTION,
+  MAX_CUSTOM_INPUT_LENGTH,
+};
