+++ src/lib/index.ts
@@ -1,1 +1,4 @@
-// lib
+import { inputContainerVariants } from "@/lib/inputContainerVariant";
+import { cn } from "@/lib/utils";
+
+export { inputContainerVariants, cn };
