+++ src/store/index.ts
@@ -1,4 +1,3 @@
-// store
 import { useToastStore } from "@/store/useToastStore";
 
 export { useToastStore };
