+++ src/types/index.ts
@@ -1,4 +1,15 @@
 import type { FooterLinkItem, FooterSnsLinkItem } from "@/types/footer-types";
 import type { ModalContextType } from "@/types/modal-context-type";
+import type { DropdownOption, SelectedOption } from "@/types/dropdown-types";
+import type { SmallToast, BigToast, Toast } from "@/types/toast-types";
 
-export type { FooterLinkItem, FooterSnsLinkItem, ModalContextType };
+export type {
+  FooterLinkItem,
+  FooterSnsLinkItem,
+  ModalContextType,
+  DropdownOption,
+  SelectedOption,
+  SmallToast,
+  BigToast,
+  Toast,
+};
